projects/llama/requirements.txt
numpy
pandas
spacy
torch
transformers<=4.45.2
jsonformer
bitsandbytes
guidance
kor
peft
trl<=0.12.0
datasets
python-dotenv
tqdm
